+++ src/fmindex-collection/search/SearchNg12.h
@@ -47,7 +47,7 @@ struct Search {
     }
 
     template <char LInfo, char RInfo>
-    void search_next(cursor_t const& cur, int e, BlockIter blockIter) const noexcept {
+    void search_next(cursor_t const& cur, size_t e, BlockIter blockIter) const noexcept {
         if (cur.count() == 0) {
             return;
         }
@@ -66,7 +66,7 @@ struct Search {
     }
 
     template <char LInfo, char RInfo, bool Right>
-    void search_next_dir(cursor_t const& cur, int e, BlockIter blockIter) const noexcept {
+    void search_next_dir(cursor_t const& cur, size_t e, BlockIter blockIter) const noexcept {
         static constexpr char TInfo = Right ? RInfo : LInfo;
 
         constexpr bool Deletion     = TInfo == 'M' or TInfo == 'D';
@@ -89,7 +89,7 @@ struct Search {
 
         if (mismatchAllowed) {
             auto cursors = std::array<cursor_t, Sigma>{};
-            for (size_t i{1}; i < Sigma; ++i) {
+            for (size_t i{1ul}; i < Sigma; ++i) {
                 if constexpr (Right) {
                     cursors[i] = cur.extendRight(i);
                 } else {
@@ -154,7 +154,7 @@ void search(index_t const & index, queries_t && queries, search_schemes_t const 
                     return s.pi[i-1] < s.pi[i]?Dir::Right:Dir::Left;
                 }
             }();
-            search2.emplace_back(Block<size_t>{{}, (size_t)s.l[i], (size_t)s.u[i], dir});
+            search2.emplace_back(Block<size_t>{{}, s.l[i], s.u[i], dir});
         }
         search_scheme2.emplace_back(move(search2));
     }
