+++ src/fmindex-collection/search/SearchNg14.h
@@ -67,7 +67,7 @@ struct Search {
 
 
     template <char LInfo, char RInfo>
-    void search_next(cursor_t const& cur, int e, BlockIter blockIter) const noexcept {
+    void search_next(cursor_t const& cur, size_t e, BlockIter blockIter) const noexcept {
         if (cur.count() == 0) {
             return;
         }
@@ -88,7 +88,7 @@ struct Search {
     }
 
     template <char LInfo, char RInfo, bool Right>
-    void search_next_dir(cursor_t const& cur, int e, BlockIter blockIter) const noexcept {
+    void search_next_dir(cursor_t const& cur, size_t e, BlockIter blockIter) const noexcept {
         static constexpr char TInfo = Right ? RInfo : LInfo;
 
         constexpr bool Deletion     = TInfo == 'M' or TInfo == 'D';
@@ -169,7 +169,7 @@ void search(index_t const & index, queries_t && queries, search_schemes_t const 
                     return s.pi[i-1] < s.pi[i]?Dir::Right:Dir::Left;
                 }
             }();
-            search2.emplace_back(Block<size_t>{{}, (size_t)s.l[i], (size_t)s.u[i], dir});
+            search2.emplace_back(Block<size_t>{{}, s.l[i], s.u[i], dir});
         }
         search_scheme2.emplace_back(move(search2));
     }
